rootdir: init.rc: Make gatord a oneshot service
A common scenario is for developers and testers to build custom kernels
for use with Android images and forget or not know to also rebuild then
install the gator module. This leads to a kernel version mismatch in the
gator module preventing it from loading and causing the user-side gator
daemon to fail to start. When this occurs, the gatord service is
repeatedly restarted every five seconds and this background activity can
affect any profiling and benchmarking activity being undertaken.

We can prevent this situation by marking gatord as a 'oneshot' service
so there is only one attempt to start it at boot.

Change-Id: I824b306f7b6618ce210bc44843dd2e1730ce8304
Signed-off-by: Jon Medhurst <tixy@linaro.org>